tty: close charshunt for bugged WebTV builds
Older WebTV builds don't send a PPP termination frame. This will listen for the flag for the end of the last PPP frame (0x7e) and the start of an AT command (0x4154). In case this might be a PPP frame start, there's no 0x4154 PPP protocol so this works out without error.